requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
httpx==0.27.2
pytest==7.4.4
pytest-asyncio==0.23.3
eval_type_backport==0.2.2
python-dotenv==1.2.1
openai-agents==0.5.0

